+++ src/magpylib_material_response/__init__.py
@@ -8,7 +8,15 @@ from __future__ import annotations
 
 from magpylib_material_response._data import get_dataset
 from magpylib_material_response.logging_config import configure_logging, disable_logging
+from magpylib_material_response.utils import from_json, to_json
 
 from ._version import version as __version__
 
-__all__ = ["__version__", "configure_logging", "disable_logging", "get_dataset"]
+__all__ = [
+    "__version__",
+    "configure_logging",
+    "disable_logging",
+    "from_json",
+    "get_dataset",
+    "to_json",
+]